delete Method

+++ Controllers/MoviesController.cs
@@ -136,6 +136,7 @@ namespace MvcMovie.Controllers
         }
 
         // GET: Movies/Delete/5
+        /*
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -151,11 +152,26 @@ namespace MvcMovie.Controllers
         }
 
         // POST: Movies/Delete/5
+        
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
             Movie movie = db.Movies.Find(id);
+            db.Movies.Remove(movie);
+            db.SaveChanges();
+            return RedirectToAction("Index");
+        }
+        */
+
+        public ActionResult Delete(FormCollection fcNotUsed, int id = 0)
+        {
+            Movie movie = db.Movies.Find(id);
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
+
             db.Movies.Remove(movie);
             db.SaveChanges();
             return RedirectToAction("Index");
